requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==7.1.2
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.62.0
h11==0.11.0
httptools==0.6.1
idna==3.6
pydantic~=1.10.14
pydantic_core==2.14.6
PyMySQL==0.10.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy~=2.0.28
starlette==0.13.6
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn~=0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0

moviepy~=1.0.3
pillow~=10.2.0
boto3~=1.34.60
botocore~=1.34.60